[helpers] Split `i18n` up into multiple modules
This aids in communicating what this module does and which areas need
unit testing.
